requirements.txt
asgiref==3.7.2
attrs==23.2.0
channels==4.0.0
click==8.1.7
Django==4.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
h11==0.14.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
psycopg==3.1.17
psycopg-binary==3.1.17
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
uvicorn==0.27.1